refactor(web): generalize enablement integration settings
